% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{backward_sampler_c}
\alias{backward_sampler_c}
\title{Backward Sampler for the Bounded Coalescent}
\usage{
backward_sampler_c(
  forward_probs,
  times,
  leaves,
  ne,
  bound,
  lineages,
  bound_size = 1L
)
}
\arguments{
\item{forward_probs}{2D array of probabilities from the forward algorithm.}

\item{times}{Vector of ordered sampling times for leaves.}

\item{leaves}{Number of leaves taken at each sampling time.}

\item{ne}{Effective population size.}

\item{bound}{Bound time.}

\item{lineages}{Vector to store sample.}

\item{bound_size}{Number of lineages at the bound (default 1).}
}
\description{
Samples the number of lineages at specified times given forward
probabilities
}
