% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoalescenceBlocking.R
\name{block_coalescences}
\alias{block_coalescences}
\title{Determine blocks of coalescent events from backward sample}
\usage{
block_coalescences(sample, times, leaves, bound)
}
\arguments{
\item{sample}{Sample of number of lineages from backward sampler}

\item{times}{Vector of ordered sampling times for leaves.}

\item{leaves}{Number of leaves taken at each sampling time.}

\item{bound}{Bound time.}

\item{ne}{Effective population size.}
}
\description{
Determine blocks of coalescent events from backward sample
}
