% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontFunctions.R
\name{bounded_forward_algorithm}
\alias{bounded_forward_algorithm}
\title{Forward algorithm for the bounded coalescent}
\usage{
bounded_forward_algorithm(t, l, ne, b)
}
\arguments{
\item{t}{Vector of leaf sampling times.}

\item{l}{vector of leaves sampled at each time.}

\item{ne}{Effective population size.}

\item{b}{Bound time.}
}
\description{
Forward algorithm for the bounded coalescent
}
