% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontFunctions.R
\name{bounded_likelihood}
\alias{bounded_likelihood}
\title{Calculate the Likelihood of Coalescence Times Under the Bounded Coalescent}
\usage{
bounded_likelihood(t, l, c, ne, b, topology = T)
}
\arguments{
\item{t}{Vector of leaf sampling times.}

\item{l}{vector of leaves sampled at each time.}

\item{c}{Vector of coalescence times.}

\item{ne}{Effective population size.}

\item{b}{Bound time.}

\item{topology}{Boolean value indicating whether the topology should be
included in the likelihood calculation.}
}
\description{
Calculate the Likelihood of Coalescence Times Under the Bounded Coalescent
}
