% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontFunctions.R
\name{bounded_likelihood_phylo}
\alias{bounded_likelihood_phylo}
\title{Calculate the likelihood of a phylogeny under the bounded coalescence}
\usage{
bounded_likelihood_phylo(phy, ne, b, topology = T)
}
\arguments{
\item{phy}{Phylogeny of class 'phylo' of 'multiPhylo'.}

\item{ne}{Effective population size.}

\item{b}{Bound time.
inputs)}

\item{topology}{Boolean value indicating whether the topology should be
included in the likelihood calculation.}
}
\description{
Calculate the likelihood of a phylogeny under the bounded coalescence
}
