% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontFunctions.R
\name{bounded_phylo_likelihood}
\alias{bounded_phylo_likelihood}
\title{Calculate the likelihood of a phylogeny under the bounded coalescence}
\usage{
bounded_phylo_likelihood(phy, ne, b, r = NA)
}
\arguments{
\item{phy}{Phylogeny of class 'phylo' or a data frame with time and ancestor
columns.}

\item{ne}{Effective population size.}

\item{b}{Bound time.}

\item{r}{Time of most recent common ancestor (needed for class 'phylo'
inputs)}
}
\description{
Calculate the likelihood of a phylogeny under the bounded coalescence
}
