% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontFunctions.R
\name{bounded_phylo_sample}
\alias{bounded_phylo_sample}
\title{Sample a phylogeny under the bounded coalescent}
\usage{
bounded_phylo_sample(t, l, ne, b, tip_labels = NA, method = "direct")
}
\arguments{
\item{t}{Vector of leaf sampling times.}

\item{l}{vector of leaves sampled at each time.}

\item{ne}{Effective population size.}

\item{b}{Bound time.}

\item{tip_labels}{Labels for sampled leaves.}

\item{method}{Sampling method for the coalescence times.}
}
\description{
Sample a phylogeny under the bounded coalescent
}
