% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontFunctions.R
\name{bounded_sample_times}
\alias{bounded_sample_times}
\title{Sample Coalescence Times Under the Bounded Coalescent}
\usage{
bounded_sample_times(t, l, ne, b, nsam = 1, method = "direct")
}
\arguments{
\item{t}{Vector of leaf sampling times.}

\item{l}{vector of leaves sampled at each time.}

\item{ne}{Effective population size.}

\item{b}{Bound time.}

\item{nsam}{Number of samples.}

\item{method}{Sampling method for the coalescence times.}
}
\description{
Sample Coalescence Times Under the Bounded Coalescent
}
