% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontFunctions.R
\name{bounded_times_likelihood}
\alias{bounded_times_likelihood}
\title{Calculate the Likelihood of Coalescence Times Under the Bounded Coalescent}
\usage{
bounded_times_likelihood(t, l, c, ne, b)
}
\arguments{
\item{t}{Vector of leaf sampling times.}

\item{l}{vector of leaves sampled at each time.}

\item{c}{Vector of coalescence times.}

\item{ne}{Effective population size.}

\item{b}{Bound time.}
}
\description{
Calculate the Likelihood of Coalescence Times Under the Bounded Coalescent
}
