% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bounded_times_likelihood_c}
\alias{bounded_times_likelihood_c}
\title{Calculate the likelihood under the bounded coalescence}
\usage{
bounded_times_likelihood_c(leaf_times, leaves, coalescence_times, ne, bound)
}
\arguments{
\item{leaf_times}{Times that leaves are added.}

\item{leaves}{Number of leaves at each time point.}

\item{coalescence_times}{Times of coalescences.}

\item{ne}{Effective population size.}

\item{bound}{Bound time.}
}
\description{
Calculate the likelihood under the bounded coalescence
}
