% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontFunctions.R
\name{coalescent_probability}
\alias{coalescent_probability}
\title{Coalescent probabilities in the homochronous setting}
\usage{
coalescent_probability(i, j, dt, ne)
}
\arguments{
\item{i}{Starting number of lineages.}

\item{j}{Ending number of lineages.}

\item{dt}{Time interval.}

\item{ne}{Effective population size.}
}
\description{
Coalescent probabilities in the homochronous setting
}
