% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constrain_coalescences_c}
\alias{constrain_coalescences_c}
\title{Constrain coalescence events to unique time intervals}
\usage{
constrain_coalescences_c(
  sample,
  times,
  leaves,
  ne,
  bound,
  const_lower,
  const_upper,
  const_lineages,
  const_events,
  norm_tol = 1e-10
)
}
\arguments{
\item{sample}{Sample of number of lineages from backward sampler}

\item{times}{Vector of ordered sampling times for leaves.}

\item{leaves}{Number of leaves taken at each sampling time.}

\item{ne}{Effective population size.}

\item{bound}{Bound time.}

\item{const_lower}{Lower bound times for each coalescence event.}

\item{const_upper}{Upper bound times for each coalescence event.}

\item{const_lineages}{Number of lineages at the start of the interval.}

\item{const_events}{Number of coalescence events in the interval.}

\item{norm_tol}{Threshold to use approximate sampling due to loss of
significance.}
}
\description{
Constrain coalescence events to unique time intervals
}
