% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_algorithm_c}
\alias{forward_algorithm_c}
\title{Forward Algorithm for the Bounded Coalescent}
\usage{
forward_algorithm_c(times, leaves, ne, bound, forward_probs)
}
\arguments{
\item{times}{Vector of ordered sampling times for leaves.}

\item{leaves}{Number of leaves taken at each sampling time.}

\item{ne}{Effective population size.}

\item{bound}{Bound time.}

\item{forward_probs}{Array to store probabilities.}
}
\description{
Calculate the forward probabilities for the bounded coalescent.
}
