% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{homochronous_probability}
\alias{homochronous_probability}
\title{Calculate coalescent probabilities (homochronous)}
\usage{
homochronous_probability(i, j, dt, Ne)
}
\arguments{
\item{i}{Integer value for the starting number of lineages.}

\item{j}{Integer value for the final number of lineages.}

\item{dt}{Time period over which coalescences can occur.}

\item{Ne}{Effective population size.}
}
\description{
Calculate the probability of i lineages coalescing down to j <= i lineages
in time dt in the homochronous setting.
}
