% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rejection_bounded_times}
\alias{rejection_bounded_times}
\title{Rejection sampling coalescence times for the bounded coalescenct}
\usage{
rejection_bounded_times(times, leaves, ne, bound, nsam = 1L)
}
\arguments{
\item{times}{Vector of ordered sampling times for leaves.}

\item{leaves}{Number of leaves taken at each sampling time.}

\item{ne}{Effective population size.}

\item{bound}{Bound time.}

\item{nsam}{Number of samples (deafault 1).}
}
\description{
Rejection sampling coalescence times for the bounded coalescenct
}
