% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_coalescence_time_c}
\alias{sample_coalescence_time_c}
\title{Sample a coalescence time within a specified interval}
\usage{
sample_coalescence_time_c(time_lower, time_upper, lineages, ne)
}
\arguments{
\item{time_lower}{Lower bound time of the interval.}

\item{time_upper}{Upper bound time of the interval.}

\item{lineages}{Starting number of lineages.}

\item{ne}{Effective population size.}
}
\description{
Sample a coalescence time within a specified interval
}
