% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_topology_c}
\alias{sample_topology_c}
\title{Sample topology given coalescent times.}
\usage{
sample_topology_c(leaf_times, leaves, coalescence_times)
}
\arguments{
\item{leaves}{Number of leaves taken at each sampling time.}

\item{forward_probs}{2D array of probabilities from the forward algorithm.}

\item{times}{Vector of ordered sampling times for leaves.}

\item{ne}{Effective population size.}

\item{bound}{Bound time.}

\item{bound_size}{Number of lineages at the bound (default 1).}
}
\description{
Sample topology given coalescent times.
}
