% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{separate_coalescences_c}
\alias{separate_coalescences_c}
\title{Separate coalescences in a block}
\usage{
separate_coalescences_c(
  coalescences,
  time_lower,
  time_upper,
  lineages_upper,
  ne
)
}
\description{
Separate coalescences in a block
}
