% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{significance_loss}
\alias{significance_loss}
\title{Check for loss of significance in proabability calculation}
\usage{
significance_loss(i, j, dt, ne)
}
\arguments{
\item{i}{Integer value for the starting number of lineages.}

\item{j}{Integer value for the final number of lineages.}

\item{dt}{Time period over which coalescences can occur.}

\item{ne}{Effective population size.}
}
\description{
Check for loss of significance in proabability calculation
}
