% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontFunctions.R
\name{bounded_sample_phylo}
\alias{bounded_sample_phylo}
\title{Sample a phylogeny under the bounded coalescent}
\usage{
bounded_sample_phylo(
  t,
  l,
  ne,
  b,
  nsam = 1,
  tip.label,
  node.label,
  method = "direct"
)
}
\arguments{
\item{t}{Vector of leaf sampling times.}

\item{l}{vector of leaves sampled at each time.}

\item{ne}{Effective population size.}

\item{b}{Bound time.}

\item{tip.label}{Labels for sampled leaves.}

\item{node.label}{Labels for nodes.}

\item{method}{Sampling method for the coalescence times.}
}
\description{
Sample a phylogeny under the bounded coalescent
}
