% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constrain_coalescences_c}
\alias{constrain_coalescences_c}
\title{Constrain coalescence events to unique time intervals}
\usage{
constrain_coalescences_c(
  sample,
  times,
  leaves,
  ne,
  bound,
  const_lower,
  const_upper,
  const_lineages,
  const_events,
  norm_tol = 1e-10
)
}
\arguments{
\item{sample}{Sample of number of lineages from backward sampler}

\item{times}{Vector of ordered sampling times for leaves.}

\item{leaves}{Number of leaves taken at each sampling time.}

\item{ne}{Effective population size.}

\item{bound}{Bound time.}
}
\description{
Constrain coalescence events to unique time intervals
}
