% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_topology_c}
\alias{sample_topology_c}
\title{Sample topology given coalescent times.}
\usage{
sample_topology_c(leaf_times, leaves, coalescence_times)
}
\arguments{
\item{leaf_times}{Vector of ordered sampling times for leaves.}

\item{leaves}{Number of leaves taken at each sampling time.}

\item{coalescence_times}{Vector of ordered coalescence times.}
}
\description{
Sample topology given coalescent times.
}
