% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{separate_coalescences_c}
\alias{separate_coalescences_c}
\title{Separate coalescences in a block}
\usage{
separate_coalescences_c(
  coalescences,
  time_lower,
  time_upper,
  lineages_upper,
  ne
)
}
\arguments{
\item{coalescences}{Number of coalescences in block.}

\item{time_lower}{Lower bound time of the block.}

\item{time_upper}{Upper bound time of the block.}

\item{lineages_upper}{Number of lineages at the start of the block.}

\item{ne}{Effective population size.}
}
\description{
Separate coalescences in a block
}
